--- a/src/index.js
+++ b/src/index.js
@@ -4,8 +4,6 @@ const hmac = require('./hmac')
 const aes = require('./aes')
 const keys = require('./keys')
 
-exports = module.exports
-
 exports.aes = aes
 exports.hmac = hmac
 exports.keys = keys