--- a/src/keys/ed25519.js
+++ b/src/keys/ed25519.js
@@ -1,51 +1,23 @@
 'use strict'
 
 const nacl = require('tweetnacl')
-const nextTick = require('async/nextTick')
 
 exports.publicKeyLength = nacl.sign.publicKeyLength
 exports.privateKeyLength = nacl.sign.secretKeyLength
 
-exports.generateKey = function (callback) {
-  nextTick(() => {
-    let result
-    try {
-      result = nacl.sign.keyPair()
-    } catch (err) {
-      return callback(err)
-    }
-    callback(null, result)
-  })
+exports.generateKey = async function () { // eslint-disable-line require-await
+  return nacl.sign.keyPair()
 }
 
 // seed should be a 32 byte uint8array
-exports.generateKeyFromSeed = function (seed, callback) {
-  nextTick(() => {
-    let result
-    try {
-      result = nacl.sign.keyPair.fromSeed(seed)
-    } catch (err) {
-      return callback(err)
-    }
-    callback(null, result)
-  })
+exports.generateKeyFromSeed = async function (seed) { // eslint-disable-line require-await
+  return nacl.sign.keyPair.fromSeed(seed)
 }
 
-exports.hashAndSign = function (key, msg, callback) {
-  nextTick(() => {
-    callback(null, Buffer.from(nacl.sign.detached(msg, key)))
-  })
+exports.hashAndSign = async function (key, msg) { // eslint-disable-line require-await
+  return Buffer.from(nacl.sign.detached(msg, key))
 }
 
-exports.hashAndVerify = function (key, sig, msg, callback) {
-  nextTick(() => {
-    let result
-    try {
-      result = nacl.sign.detached.verify(msg, sig, key)
-    } catch (err) {
-      return callback(err)
-    }
-
-    callback(null, result)
-  })
+exports.hashAndVerify = async function (key, sig, msg) { // eslint-disable-line require-await
+  return nacl.sign.detached.verify(msg, sig, key)
 }