--- a/src/keys/key-stretcher.js
+++ b/src/keys/key-stretcher.js
@@ -1,6 +1,5 @@
 'use strict'
 
-const whilst = require('async/whilst')
 const hmac = require('../hmac')
 
 const cipherMap = {
@@ -20,15 +19,15 @@ const cipherMap = {
 
 // Generates a set of keys for each party by stretching the shared key.
 // (myIV, theirIV, myCipherKey, theirCipherKey, myMACKey, theirMACKey)
-module.exports = (cipherType, hash, secret, callback) => {
+module.exports = async (cipherType, hash, secret) => {
   const cipher = cipherMap[cipherType]
 
   if (!cipher) {
-    return callback(new Error('unkown cipherType passed'))
+    throw new Error('unkown cipherType passed')
   }
 
   if (!hash) {
-    return callback(new Error('unkown hashType passed'))
+    throw new Error('unkown hashType passed')
   }
 
   const cipherKeySize = cipher.keySize
@@ -37,72 +36,38 @@ module.exports = (cipherType, hash, secret, callback) => {
   const seed = Buffer.from('key expansion')
   const resultLength = 2 * (ivSize + cipherKeySize + hmacKeySize)
 
-  hmac.create(hash, secret, (err, m) => {
-    if (err) {
-      return callback(err)
+  const m = await hmac.create(hash, secret)
+  let a = await m.digest(seed)
+
+  let result = []
+  let j = 0
+
+  while (j < resultLength) {
+    const b = await m.digest(Buffer.concat([a, seed]))
+    let todo = b.length
+
+    if (j + todo > resultLength) {
+      todo = resultLength - j
     }
 
-    m.digest(seed, (err, a) => {
-      if (err) {
-        return callback(err)
-      }
+    result.push(b)
+    j += todo
+    a = await m.digest(a)
+  }
 
-      let result = []
-      let j = 0
+  const half = resultLength / 2
+  const resultBuffer = Buffer.concat(result)
+  const r1 = resultBuffer.slice(0, half)
+  const r2 = resultBuffer.slice(half, resultLength)
 
-      whilst(
-        () => j < resultLength,
-        stretch,
-        finish
-      )
-
-      function stretch (cb) {
-        m.digest(Buffer.concat([a, seed]), (err, b) => {
-          if (err) {
-            return cb(err)
-          }
-
-          let todo = b.length
-
-          if (j + todo > resultLength) {
-            todo = resultLength - j
-          }
-
-          result.push(b)
-
-          j += todo
-
-          m.digest(a, (err, _a) => {
-            if (err) {
-              return cb(err)
-            }
-            a = _a
-            cb()
-          })
-        })
-      }
-
-      function finish (err) {
-        if (err) {
-          return callback(err)
-        }
-
-        const half = resultLength / 2
-        const resultBuffer = Buffer.concat(result)
-        const r1 = resultBuffer.slice(0, half)
-        const r2 = resultBuffer.slice(half, resultLength)
-
-        const createKey = (res) => ({
-          iv: res.slice(0, ivSize),
-          cipherKey: res.slice(ivSize, ivSize + cipherKeySize),
-          macKey: res.slice(ivSize + cipherKeySize)
-        })
-
-        callback(null, {
-          k1: createKey(r1),
-          k2: createKey(r2)
-        })
-      }
-    })
+  const createKey = (res) => ({
+    iv: res.slice(0, ivSize),
+    cipherKey: res.slice(ivSize, ivSize + cipherKeySize),
+    macKey: res.slice(ivSize + cipherKeySize)
   })
+
+  return {
+    k1: createKey(r1),
+    k2: createKey(r2)
+  }
 }