--- a/src/keys/rsa-browser.js
+++ b/src/keys/rsa-browser.js
@@ -1,13 +1,12 @@
 'use strict'
 
-const nodeify = require('../nodeify')
-const webcrypto = require('../webcrypto')
+const webcrypto = require('../webcrypto.js')
 const randomBytes = require('../random-bytes')
 
 exports.utils = require('./rsa-utils')
 
-exports.generateKey = function (bits, callback) {
-  nodeify(webcrypto.subtle.generateKey(
+exports.generateKey = async function (bits) {
+  const pair = await webcrypto.subtle.generateKey(
     {
       name: 'RSASSA-PKCS1-v1_5',
       modulusLength: bits,
@@ -17,16 +16,18 @@ exports.generateKey = function (bits, callback) {
     true,
     ['sign', 'verify']
   )
-    .then(exportKey)
-    .then((keys) => ({
-      privateKey: keys[0],
-      publicKey: keys[1]
-    })), callback)
+
+  const keys = await exportKey(pair)
+
+  return {
+    privateKey: keys[0],
+    publicKey: keys[1]
+  }
 }
 
 // Takes a jwk key
-exports.unmarshalPrivateKey = function (key, callback) {
-  const privateKey = webcrypto.subtle.importKey(
+exports.unmarshalPrivateKey = async function (key) {
+  const privateKey = await webcrypto.subtle.importKey(
     'jwk',
     key,
     {
@@ -37,22 +38,26 @@ exports.unmarshalPrivateKey = function (key, callback) {
     ['sign']
   )
 
-  nodeify(Promise.all([
+  const pair = [
     privateKey,
-    derivePublicFromPrivate(key)
-  ]).then((keys) => exportKey({
+    await derivePublicFromPrivate(key)
+  ]
+
+  const keys = await exportKey({
+    privateKey: pair[0],
+    publicKey: pair[1]
+  })
+
+  return {
     privateKey: keys[0],
     publicKey: keys[1]
-  })).then((keys) => ({
-    privateKey: keys[0],
-    publicKey: keys[1]
-  })), callback)
+  }
 }
 
 exports.getRandomValues = randomBytes
 
-exports.hashAndSign = function (key, msg, callback) {
-  nodeify(webcrypto.subtle.importKey(
+exports.hashAndSign = async function (key, msg) {
+  const privateKey = await webcrypto.subtle.importKey(
     'jwk',
     key,
     {
@@ -61,17 +66,19 @@ exports.hashAndSign = function (key, msg, callback) {
     },
     false,
     ['sign']
-  ).then((privateKey) => {
-    return webcrypto.subtle.sign(
-      { name: 'RSASSA-PKCS1-v1_5' },
-      privateKey,
-      Uint8Array.from(msg)
-    )
-  }).then((sig) => Buffer.from(sig)), callback)
+  )
+
+  const sig = await webcrypto.subtle.sign(
+    { name: 'RSASSA-PKCS1-v1_5' },
+    privateKey,
+    Uint8Array.from(msg)
+  )
+
+  return Buffer.from(sig)
 }
 
-exports.hashAndVerify = function (key, sig, msg, callback) {
-  nodeify(webcrypto.subtle.importKey(
+exports.hashAndVerify = async function (key, sig, msg) {
+  const publicKey = await webcrypto.subtle.importKey(
     'jwk',
     key,
     {
@@ -80,14 +87,14 @@ exports.hashAndVerify = function (key, sig, msg, callback) {
     },
     false,
     ['verify']
-  ).then((publicKey) => {
-    return webcrypto.subtle.verify(
-      { name: 'RSASSA-PKCS1-v1_5' },
-      publicKey,
-      sig,
-      msg
-    )
-  }), callback)
+  )
+
+  return webcrypto.subtle.verify(
+    { name: 'RSASSA-PKCS1-v1_5' },
+    publicKey,
+    sig,
+    msg
+  )
 }
 
 function exportKey (pair) {