@alanshaw 2018-10-28
@mkg20001 any idea why `ursa-optional` is failing to install on windows now? We get no error message :(