@mkg20001 2018-10-28
@alanshaw The error messages usually get stored in `node_modules/ursa-optional/stderr.log` and `node_modules/ursa-optional/stdout.log`.
I'd love to debug that but sadly I'm stuck at a place with an internet connection so slow it takes _up to 10 seconds_ to SSH into anything and the only windows machine I have is a VM on a server.
At least ursa-optional is optional so it's not that much of a problem.