@alanshaw 2018-10-29
Ah, thanks...we can't look at them in CI though. Is there a reason why they're not just sent to stdout and stderr? Makes for nice npm install output but difficult to debug from here.