@mkg20001 2018-10-29
@coolaj86 suggested to add that as people might see the error message and think the thing is broken while in fact it's not. I'll release a new version with an env flag `URSA_VERBOSE` which would print the output directly if set, that would have to be enabled on CI, too