@diasdavid 2018-11-11
> This is largely because the Node.js crypto API is synchronous. However the webcrypto API is async so I assume this has been done to present a consistent interface to developers in both Node.js and the browser. Maybe someone can verify that for me?

This is exactly correct. We need to keep things `async` otherwise we would have to interact with both the Node.js Crypto and WebCrypto differently changing a bunch of modules upstream.