@jacobheun 2019-07-10
@dignifiedquire I am going to merge this in so that we can incorporate #153 as well before the release. If there are any issues you see with this we can do a follow up PR before the release.